# server/openapi/desktop-client-v1.yaml
openapi: 3.1.0
info:
  title: ClawHub Desktop Client API
  version: 1.2.0
  description: |
    Versioned contract for Agent appearance, paid hiring, account-synchronized
    desktop sidebar preferences, and the desktop template catalog. Offer is an
    immutable receipt created only after verified payment succeeds.
servers:
  - url: https://www.iclawstore.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/ai-direct-hiring/paid-hiring/orders:
    post:
      tags: [Offers]
      summary: Create an idempotent Alipay hiring order from the server-side price
      operationId: createPaidHiringOrder
      parameters:
        - {
            name: Idempotency-Key,
            in: header,
            required: true,
            schema: { type: string, maxLength: 128 },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [companyId, roleId, positionId, agentId]
              properties:
                companyId: { type: string, maxLength: 36 }
                projectId: { type: [string, "null"], maxLength: 36 }
                roleId: { type: string, maxLength: 36 }
                positionId: { type: string, maxLength: 36 }
                agentId: { type: string, maxLength: 36 }
      responses:
        "201": { description: Pending Alipay order with a server-signed payUrl }
        "200": { description: Replay of the same idempotent pending order }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/ForbiddenScope" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { description: Alipay merchant configuration is unavailable }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          enum:
            [
              AUTH_REQUIRED,
              FORBIDDEN_SCOPE,
              NOT_FOUND,
              VALIDATION_ERROR,
              INVALID_TRANSITION,
              RUN_NOT_RECOVERABLE,
              RUNTIME_CAPABILITY_DISABLED,
              APPEARANCE_CONTROL_CONFLICT,
              ASSET_LIMIT_EXCEEDED,
              ASSET_TOO_LARGE,
              UNSUPPORTED_MEDIA_TYPE,
              ASSET_IN_USE,
              REVISION_CONFLICT,
              PRECONDITION_REQUIRED,
              TEMPLATE_ENTITLEMENT_REQUIRED,
              TEMPLATE_NOT_INSTALLABLE,
              DUPLICATE_ENTRY,
              INTERNAL_ERROR,
            ]
        error: { type: string }
        details: {}
...
  responses:
    AuthRequired:
      description: Missing, invalid, or expired JWT
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ForbiddenScope:
      description: The authenticated user lacks the required resource scope
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: Request validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: State, ownership, uniqueness, limit, or in-use conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...